requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
streamlit>=1.38.0
pymupdf>=1.24.0
python-dotenv>=1.0.0
